test(omo-ai): split senpi component shim coverage
